+++ test/test_helper.rb
@@ -9,6 +9,8 @@ require "minitest/autorun"
 class Minitest::Test
   protected
 
+  make_my_diffs_pretty!
+
   def thread(&blk)
     Thread.new(&blk).tap do |t|
       t.abort_on_exception = true