+++ test/unit/simulates_argument_error/reconciles_args_with_params_test.rb
@@ -1,0 +1,20 @@
+require "test_helper"
+
+module Mocktail
+  class ReconcilesArgsWithParamsTest < Minitest::Test
+    def setup
+      @subject = ReconcilesArgsWithParams.new
+    end
+
+    def test_ensure_unknown_keyword_fails
+      assert_equal false, @subject.reconcile(
+        Signature.new(
+          positional_params: Params.new(all: []),
+          positional_args: [],
+          keyword_params: Params.new(all: [:a], optional: [:a]),
+          keyword_args: {b: 42}
+        )
+      )
+    end
+  end
+end